refactor: add dashboardview layout
