+++ pages/profile/delete-account.tsx
@@ -1,9 +1,9 @@
 import Head from 'next/head';
 import React, { ReactElement } from 'react';
 import UserLayout from '../../src/features/common/Layout/UserLayout/UserLayout';
-import DeleteProfile from '../../src/features/user/Settings/DeleteProfile';
 import { useTranslation } from 'next-i18next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import DeleteProfile from '../../src/features/user/Settings/DeleteProfile';
 
 function DeleteProfilePage(): ReactElement {
   const { t } = useTranslation('me');
