+++ src/features/user/Settings/DeleteProfile/DeleteProfile.module.scss
@@ -1,4 +1,5 @@
-@import './../../../theme/theme';
+@import '../../../../theme/theme';
+
 
 .deleteModal {
   display: flex;
